requirements.txt
python-dotenv>=1.0.0
requests>=2.25.0
gradio>=3.32.0
pillow>=9.0.0
typing-extensions>=4.0.0